requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2023.3
psycopg2-binary==2.9.9
flask==3.0.0
gunicorn==21.2.0
asgiref==3.7.2

